refactor(media): extract WSConnection from WebsocketTransport
